pyproject.toml
[project]
name = "object-harvest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"pydantic>=2.7",
	"pillow>=10.0",
	"numpy>=1.26",
	"tqdm>=4.66",
	"openai>=1.0.0",
	"python-dotenv>=1.0.0",
]

[project.optional-dependencies]
hf = ["datasets>=2.19.0", "hf-transfer"]
openai = ["openai>=1.0.0"]
all = ["datasets>=2.19.0", "openai>=1.0.0", "hf-transfer"]

[project.scripts]
object-harvest = "object_harvest.cli:main"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "S", "W"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
